Billing/app/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyMySQL==1.1.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
